requirements.txt
# Streamlit Web App
streamlit

# MySQL Database Connection
mysql-connector-python

# Machine Learning Model
scikit-learn
joblib

# Image Handling
Pillow

# For nicer UI components (sliders, tables)
streamlit-aggrid

# pip install -r requirements.txt
